modify

+++ controllers/AboutController.php
@@ -48,12 +48,15 @@ class AboutController extends \yii\web\Controller
 
     public function actionUpdate()
     {
-        $post = \Yii::$app->request->post();
-        $id = isset($post['id']) ? $post['id'] : '';
-        $title = isset($post['title']) ? $post['title'] : '';
-        $content = isset($post['content']) ? $post['content'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $id = isset($post['id']) ? $post['id'] : '';
+            $title = isset($post['title']) ? $post['title'] : '';
+            $content = isset($post['content']) ? $post['content'] : '';
 
-        $res = $this->getModel()->updateOne($id, $title, $content);
+            $res = $this->getModel()->updateOne($id, $title, $content);
+        }
 
         if(!$res) return $this->fail();
         return $this->succeed();
@@ -61,11 +64,14 @@ class AboutController extends \yii\web\Controller
 
     public function actionInsert()
     {
-        $post = \Yii::$app->request->post();
-        $title = isset($post['title']) ? $post['title'] : '';
-        $content = isset($post['content']) ? $post['content'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $title = isset($post['title']) ? $post['title'] : '';
+            $content = isset($post['content']) ? $post['content'] : '';
 
-        $res = $this->getModel()->insertOne($title, $content);
+            $res = $this->getModel()->insertOne($title, $content);
+        }
 
         if(!$res) return $this->fail();
         return $this->succeed(array('insert_id' => $res));
@@ -73,9 +79,12 @@ class AboutController extends \yii\web\Controller
 
     public function actionDelete()
     {
-        $post = \Yii::$app->request->post();
-        $id = isset($post['id']) ? $post['id'] : '';
-        $res = $this->getModel()->deleteOne($id);
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $id = isset($post['id']) ? $post['id'] : '';
+            $res = $this->getModel()->deleteOne($id);
+        }
 
         if(!$res) return $this->fail();
         return $this->succeed();
