+++ controllers/CollegeController.php
@@ -48,39 +48,51 @@ class CollegeController extends \yii\web\Controller
 
     public function actionUpdate()
     {
-        $post = \Yii::$app->request->post();
-        $id = isset($post['id']) ? $post['id'] : '';
-        $title = isset($post['title']) ? $post['title'] : '';
-        $photo = isset($post['photo']) ? $post['photo'] : '';
-        $name = isset($post['name']) ? $post['name'] : '';
-        $description = isset($post['description']) ? $post['description'] : '';
-        $price = isset($post['price']) ? $post['price'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $id = isset($post['id']) ? $post['id'] : '';
+            $title = isset($post['title']) ? $post['title'] : '';
+            $photo = isset($post['photo']) ? $post['photo'] : '';
+            $name = isset($post['name']) ? $post['name'] : '';
+            $description = isset($post['description']) ? $post['description'] : '';
+            $price = isset($post['price']) ? $post['price'] : '';
 
-        $res = $this->getModel()->updateOne($id, $title, $photo, $name, $description, $price);
+            $res = $this->getModel()->updateOne($id, $title, $photo, $name, $description, $price);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed();
     }
 
     public function actionInsert()
     {
-        $post = \Yii::$app->request->post();
-        $title = isset($post['title']) ? $post['title'] : '';
-        $photo = isset($post['photo']) ? $post['photo'] : '';
-        $name = isset($post['name']) ? $post['name'] : '';
-        $description = isset($post['description']) ? $post['description'] : '';
-        $price = isset($post['price']) ? $post['price'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $title = isset($post['title']) ? $post['title'] : '';
+            $photo = isset($post['photo']) ? $post['photo'] : '';
+            $name = isset($post['name']) ? $post['name'] : '';
+            $description = isset($post['description']) ? $post['description'] : '';
+            $price = isset($post['price']) ? $post['price'] : '';
 
-        $res = $this->getModel()->insertOne($title, $photo, $name, $description, $price);
+            $res = $this->getModel()->insertOne($title, $photo, $name, $description, $price);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed(array('insert_id' => $res));
     }
 
     public function actionDelete()
     {
-        $post = \Yii::$app->request->post();
-        $id = isset($post['id']) ? $post['id'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $id = isset($post['id']) ? $post['id'] : '';
 
-        $res = $this->getModel()->deleteOne($id);
+            $res = $this->getModel()->deleteOne($id);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed();
     }
