+++ controllers/CustomersController.php
@@ -48,37 +48,49 @@ class CustomersController extends \yii\web\Controller
 
     public function actionUpdate()
     {
-        $post = \Yii::$app->request->post();
-        $id = isset($post['id']) ? $post['id'] : '';
-        $username = isset($post['username']) ? $post['username'] : '';
-        $password = isset($post['password']) ? $post['password'] : '';
-        $email = isset($post['email']) ? $post['email'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $id = isset($post['id']) ? $post['id'] : '';
+            $username = isset($post['username']) ? $post['username'] : '';
+            $password = isset($post['password']) ? $post['password'] : '';
+            $email = isset($post['email']) ? $post['email'] : '';
 
-        $password = MD5($password);
-        $res = $this->getModel()->updateOne($id, $username, $password, $email);
+            $password = MD5($password);
+            $res = $this->getModel()->updateOne($id, $username, $password, $email);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed();
     }
 
     public function actionInsert()
     {
-        $post = \Yii::$app->request->post();
-        $username = isset($post['username']) ? $post['username'] : '';
-        $password = isset($post['password']) ? $post['password'] : '';
-        $email = isset($post['email']) ? $post['email'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $username = isset($post['username']) ? $post['username'] : '';
+            $password = isset($post['password']) ? $post['password'] : '';
+            $email = isset($post['email']) ? $post['email'] : '';
 
-        $password = MD5($password);
-        $res = $this->getModel()->insertOne($username, $password, $email);
+            $password = MD5($password);
+            $res = $this->getModel()->insertOne($username, $password, $email);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed(array('insert_id' => $res));
     }
 
     public function actionDelete()
     {
-        $post = \Yii::$app->request->post();
-        $id = isset($post['id']) ? $post['id'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $id = isset($post['id']) ? $post['id'] : '';
 
-        $res = $this->getModel()->deleteOne($id);
+            $res = $this->getModel()->deleteOne($id);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed();
     }
