+++ controllers/HelpController.php
@@ -48,35 +48,47 @@ class HelpController extends \yii\web\Controller
 
     public function actionUpdate()
     {
-        $post = \Yii::$app->request->post();
-        $id = isset($post['id']) ? $post['id'] : '';
-        $type = isset($post['type']) ? $post['type'] : '';
-        $title = isset($post['title']) ? $post['title'] : '';
-        $content = isset($post['content']) ? $post['content'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $id = isset($post['id']) ? $post['id'] : '';
+            $type = isset($post['type']) ? $post['type'] : '';
+            $title = isset($post['title']) ? $post['title'] : '';
+            $content = isset($post['content']) ? $post['content'] : '';
 
-        $res = $this->getModel()->updateOne($id, $type, $title, $content);
+            $res = $this->getModel()->updateOne($id, $type, $title, $content);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed();
     }
 
     public function actionInsert()
     {
-        $post = \Yii::$app->request->post();
-        $type = isset($post['type']) ? $post['type'] : '';
-        $title = isset($post['title']) ? $post['title'] : '';
-        $content = isset($post['content']) ? $post['content'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $type = isset($post['type']) ? $post['type'] : '';
+            $title = isset($post['title']) ? $post['title'] : '';
+            $content = isset($post['content']) ? $post['content'] : '';
 
-        $res = $this->getModel()->insertOne($type, $title, $content);
+            $res = $this->getModel()->insertOne($type, $title, $content);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed(array('insert_id' => $res));
     }
 
     public function actionDelete()
     {
-        $post = \Yii::$app->request->post();
-        $id = isset($post['id']) ? $post['id'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $id = isset($post['id']) ? $post['id'] : '';
 
-        $res = $this->getModel()->deleteOne($id);
+            $res = $this->getModel()->deleteOne($id);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed();
     }
