+++ controllers/PartnerController.php
@@ -48,34 +48,46 @@ class PartnerController extends \yii\web\Controller
 
     public function actionUpdate()
     {
-        $post = \Yii::$app->request->post();
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
 
-        $id = isset($post['id']) ? $post['id'] : '';
-        $photo = isset($post['photo']) ? $post['photo'] : '';
-        $order = isset($post['order']) ? $post['order'] : 0;
+            $id = isset($post['id']) ? $post['id'] : '';
+            $photo = isset($post['photo']) ? $post['photo'] : '';
+            $order = isset($post['order']) ? $post['order'] : 0;
 
-        $res = $this->getModel()->updateOne($id, $photo, $order);
+            $res = $this->getModel()->updateOne($id, $photo, $order);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed();
     }
 
     public function actionInsert()
     {
-        $post = \Yii::$app->request->post();
-        $photo = isset($post['photo']) ? $post['photo'] : '';
-        $order = isset($post['order']) ? $post['order'] : 0;
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $photo = isset($post['photo']) ? $post['photo'] : '';
+            $order = isset($post['order']) ? $post['order'] : 0;
 
-        $res = $this->getModel()->insertOne($photo, $order);
+            $res = $this->getModel()->insertOne($photo, $order);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed(array('insert_id' => $res));
     }
 
     public function actionDelete()
     {
-        $post = \Yii::$app->request->post();
-        $id = isset($post['id']) ? $post['id'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $id = isset($post['id']) ? $post['id'] : '';
 
-        $res = $this->getModel()->deleteOne($id);
+            $res = $this->getModel()->deleteOne($id);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed();
     }
