+++ controllers/ProductMainController.php
@@ -48,37 +48,49 @@ class ProductMainController extends \yii\web\Controller
 
     public function actionUpdate()
     {
-        $post = \Yii::$app->request->post();
-        $id = isset($post['id']) ? $post['id'] : '';
-        $content = isset($post['content']) ? $post['content'] : '';
-        $photo = isset($post['photo']) ? $post['photo'] : '';
-        $link_name = isset($post['link_name']) ? $post['link_name'] : '';
-        $link_url = isset($post['link_url']) ? $post['link_url'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $id = isset($post['id']) ? $post['id'] : '';
+            $content = isset($post['content']) ? $post['content'] : '';
+            $photo = isset($post['photo']) ? $post['photo'] : '';
+            $link_name = isset($post['link_name']) ? $post['link_name'] : '';
+            $link_url = isset($post['link_url']) ? $post['link_url'] : '';
 
-        $res = $this->getModel()->updateOne($id, $content, $photo, $link_name, $link_url);
+            $res = $this->getModel()->updateOne($id, $content, $photo, $link_name, $link_url);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed();
     }
 
     public function actionInsert()
     {
-        $post = \Yii::$app->request->post();
-        $content = isset($post['content']) ? $post['content'] : '';
-        $photo = isset($post['photo']) ? $post['photo'] : '';
-        $link_name = isset($post['link_name']) ? $post['link_name'] : '';
-        $link_url = isset($post['link_url']) ? $post['link_url'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $content = isset($post['content']) ? $post['content'] : '';
+            $photo = isset($post['photo']) ? $post['photo'] : '';
+            $link_name = isset($post['link_name']) ? $post['link_name'] : '';
+            $link_url = isset($post['link_url']) ? $post['link_url'] : '';
 
-        $res = $this->getModel()->insertOne($content, $photo, $link_name, $link_url);
+            $res = $this->getModel()->insertOne($content, $photo, $link_name, $link_url);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed(array('insert_id' => $res));
     }
 
     public function actionDelete()
     {
-        $post = \Yii::$app->request->post();
-        $id = isset($post['id']) ? $post['id'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $id = isset($post['id']) ? $post['id'] : '';
 
-        $res = $this->getModel()->deleteOne($id);
+            $res = $this->getModel()->deleteOne($id);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed();
     }
