+++ controllers/SiteMainController.php
@@ -49,35 +49,47 @@ class SiteMainController extends \yii\web\Controller
 
     public function actionUpdate()
     {
-        $post = \Yii::$app->request->post();
-        $id = isset($post['id']) ? $post['id'] : '';
-        $title = isset($post['title']) ? $post['title'] : '';
-        $content = isset($post['content']) ? $post['content'] : '';
-        $photo = isset($post['photo']) ? $post['photo'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $id = isset($post['id']) ? $post['id'] : '';
+            $title = isset($post['title']) ? $post['title'] : '';
+            $content = isset($post['content']) ? $post['content'] : '';
+            $photo = isset($post['photo']) ? $post['photo'] : '';
 
-        $res = $this->getModel()->updateOne($id, $title, $content, $photo);
+            $res = $this->getModel()->updateOne($id, $title, $content, $photo);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed();
     }
 
     public function actionInsert()
     {
-        $post = \Yii::$app->request->post();
-        $title = isset($post['title']) ? $post['title'] : '';
-        $content = isset($post['content']) ? $post['content'] : '';
-        $photo = isset($post['photo']) ? $post['photo'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $title = isset($post['title']) ? $post['title'] : '';
+            $content = isset($post['content']) ? $post['content'] : '';
+            $photo = isset($post['photo']) ? $post['photo'] : '';
 
-        $res = $this->getModel()->insertOne($title, $content, $photo);
+            $res = $this->getModel()->insertOne($title, $content, $photo);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed(array('insert_id' => $res));
     }
 
     public function actionDelete()
     {
-        $post = \Yii::$app->request->post();
-        $id = isset($post['id']) ? $post['id'] : '';
+        $res = false;
+        if (\Yii::$app->request->isPost) {
+            $post = \Yii::$app->request->post();
+            $id = isset($post['id']) ? $post['id'] : '';
 
-        $res = $this->getModel()->deleteOne($id);
+            $res = $this->getModel()->deleteOne($id);
+        }
+
         if(!$res) return $this->fail();
         return $this->succeed();
     }
